Update example programs
