+++ Makefile
@@ -1,5 +1,6 @@
 default_target: all
-all: precision benchmarking errors
+.PHONY: all precision benchmark errors
+all: precision benchmark errors
 
 CXXFLAGS = -std=c++14 -I./src/ -Wall
 
@@ -7,9 +8,9 @@ precision:
 	@echo Compiling \"precision\" example program ...
 	@g++ examples/precision.cpp ${CXXFLAGS} -o ./precision
 
-benchmarking:
-	@echo Compiling \"benchmarking\" example program ...
-	@g++ examples/benchmarking.cpp ${CXXFLAGS} -o ./benchmarking
+benchmark:
+	@echo Compiling \"benchmark\" example program ...
+	@g++ examples/benchmark.cpp ${CXXFLAGS} -o ./benchmark
 
 errors:
 	@echo Compiling \"errors\" example program ...
