+++ examples/precision.cpp
@@ -1,14 +1,19 @@
+
+///
+/// @file precision.cpp Example program for precision testing.
+///
+
 #include "chebyshev.h"
 using namespace ch;
 
 
-Real f(Real x) {
-	return x * x;
+double f(double x) {
+	return x * std::sqrt(x);
 }
 
 
-Real f_a(Real x) {
-	return x * x + sqrt(x) * 0.1 + 0.001;
+double f_a(double x) {
+	return x * std::sqrt(x + 1E-12);
 }
 
 
@@ -17,45 +22,12 @@ int main(int argc, char const *argv[]) {
 	// Setup the precision testing environment
 	prec::setup("chebyshev", argc, argv);
 
-		// Estimate errors on f_a on [0, 10]
-		prec::estimate("f_a", f_a, f, interval(0, 10));
+		// Estimate errors on f_a on [0, 1000]
+		prec::estimate("f_a", f_a, f, prec::interval(0, 100));
 
-		// Estimate errors on f_a on multiple intervals
-		prec::estimate("f_a", f_a, f, {
-			interval(0, 1),
-			interval(1, 10),
-			interval(10, 100),
-			interval(100, 1000)
-		});
-
-		// Check that two values are almost equal
+		// Check that two values are equal
+		// up to a tolerance
 		prec::equals("f_a", f_a(1), 1, 0.2);
-
-		// Check multiple pairs of values
-		prec::equals("f_a", {
-			{f_a(1), 1},
-			{f_a(2), 4},
-			{f_a(3), 9}
-		});
-
-		// Or more simply, if the function is the same:
-		prec::equals("f_a", f_a, {
-			{1, 1},
-			{2, 4},
-			{3, 9},
-			{4, 16}
-		});
-
-		// Custom precision estimates are also supported
-		prec::estimate("custom",
-			[](interval k, Real tol, uint32_t iter){
-				prec::estimate_result r;
-				r.maxErr = tol / 2.0;
-				return r;
-			}, {
-				interval(0, 1),
-				interval(2, 3)
-			});
 
 	// Stop precision testing
 	prec::terminate();
